% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candela.R
\name{candela}
\alias{candela}
\title{Candela}
\usage{
candela(name, width = NULL, height = NULL, ...)
}
\arguments{
\item{name}{The name of the Candela component to initialize.
See details.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{\ldots}{Additional parameters for the specific Candela component.
See details.}
}
\description{
Interoperable web visualization components.
}
\details{
See \url{https://candela.readthedocs.io/en/latest/#components} for
  component list and parameters.
}

